fix(tool): split csr toggle fields below the max-lines limit
The combined `csr-key-source-fields.tsx` reached 318 effective lines
after extracting each selector into its own subcomponent. Move the
five toggle-based subcomponents (KeySourceField, AlgorithmField,
RsaKeySizeField, RsaHashField, EcCurveField) into a sibling
`csr-toggle-fields.tsx` so each file stays under the 300-line cap.
